refactor(serve): extract resolveBindHost to its own module (#616)
server.ts has a module-level auto-start side effect (`if
(!process.env.MAW_CLI) startServer()`), which means importing it in
a test process competes for port 3456 with other suites. Move the
pure heuristic into src/core/bind-host.ts so tests can import it
without triggering that side effect.

No behavior change — server.ts now imports and delegates.
